requirements.txt
typer>=0.9.0
rich>=13.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6
httpx>=0.25.0
openai>=1.6.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
tenacity>=8.2.0

# Dev/test
pytest>=7.4.0
pytest-asyncio>=0.23.0
